+++ requests_magic/request.py
@@ -3,29 +3,33 @@
 Attributes:
     default_headers (dict): 当请求没有指定 headers 时，会从这里copy一份，默认是空的。
 """
-import json
-from typing import Callable
-from .logger import Logger
-from .middleware import *
 import threading
+import json
 import time
 import hashlib
-import requests_magic.utils as utils
+from typing import Callable, NoReturn, Dict, Union
+from .logger import Logger
+from .middleware import *
+from .utils import HasNameObject, getattr_in_module
+
+__FUCK_CIRCULAR_IMPORT = False
+if __FUCK_CIRCULAR_IMPORT:
+    from .scheduler import Scheduler
 
 default_headers: dict = {}
 
 
-class Request:
+class Request(HasNameObject):
     """表示一个请求，这里会自动创建 RequestThread
     """
 
-    def __init__(self, url: str, callback: Callable,
+    def __init__(self, url: str, callback: Callable[[Any, 'Request'], NoReturn],
                  data: dict = None, method: str = 'GET', headers: dict = None,
                  time_out: int = 10, time_out_wait: int = 15, time_out_retry: int = 3,
                  tags: dict = None,
-                 downloader: Callable = requests_downloader,
-                 downloader_filter: Callable = requests_downloader_filter,
-                 preparse: Callable = None,
+                 downloader: Callable[['Request'], NoReturn] = requests_downloader,
+                 downloader_filter: Callable[['Result', 'Request'], NoReturn] = requests_downloader_filter,
+                 preparse: Callable[['Result', 'Request'], NoReturn] = None,
                  name: str = '',
                  **kwargs):
         """表示一个请求，这里会自动创建 RequestThread
@@ -40,9 +44,9 @@ class Request:
             time_out_wait: 超时后重试前的等待时间，默认：15秒
             time_out_retry: 超时重试次数，默认：3次
             tags: 标签，用来记录一些额外内容，可以用来在解析函数、下载中间件等东西之间传递信息
-            downloader: 下载器，这个方法需要有一个参数表示 Request 并返回请求结果，默认使用基于 requests 实现
-            downloader_filter: 下载过滤器，这个方法需要有两个参数分别表示 请求结果 和 Request，默认使用基于 requests 实现
-            preparse: 预解析器，默认使用解析函数所在爬虫类的 parparse 方法
+            downloader: 下载器，这个方法需要有一个参数表示 Request 并返回请求结果，默认使用基于 requests 实现，如果要持久化请求，则需要把你的下载器定义在某个模块顶级
+            downloader_filter: 下载过滤器，这个方法需要有两个参数分别表示 请求结果 和 Request，默认使用基于 requests 实现，如果要持久化请求，则需要把你的下载过滤器定义在某个模块顶级
+            preparse: 预解析器，这必须是爬虫类中的方法，默认使用解析函数所在爬虫类的 preparse 方法
             name: 请求的名字，希望能帮助 debug
             kwargs: 直接记录在自己的 kwargs 属性上，默认的 requests 下载器会把这里的值添加到 requests.request 方法的参数上
         """
@@ -61,7 +65,7 @@ class Request:
         self.method: str = method
         self.headers: dict = headers
 
-        self.callback: Callable = callback
+        self.callback: Callable[[Any, 'Request'], NoReturn] = callback
         self.downloader = downloader
         self.downloader_filter = downloader_filter
         self.tags: dict = tags
@@ -89,8 +93,15 @@ class Request:
 
         self.thread: RequestThread = None
 
-    def __str__(self) -> str:
-        return f"[Request - {self.name}]" if self.name else ''
+    def is_requesting(self) -> bool:
+        """ 是否正在请求中，根据是否存在下载线程判断
+        """
+        return self.thread is not None
+
+    def is_finish(self) -> bool:
+        """ 是否已经请求完成，根据是否有结果和 is_requesting 判断
+        """
+        return self.result and not self.is_requesting()
 
     def start(self):
         """开始下载，自动创建 RequestThread，下载完成后会自动调用调度器的方法
@@ -119,32 +130,31 @@ class Request:
         self.start_time = -1
         self.total_time = -1
         self.result = None
-        self.stop_thread()
+        self.stop()
         self.scheduler.downloader_retry(self)
 
     def abandon(self) -> None:
         """放弃请求
         """
-        self.stop_thread()
+        self.stop()
         self.scheduler.downloader_abandon(self)
 
     def finish(self) -> None:
         """完成请求
         """
-        self.stop_thread()
+        self.stop()
         self.scheduler.downloader_finish(self.result, self)
 
-    def stop_thread(self) -> None:
+    def stop(self) -> None:
         """终止请求线程
         """
-        del self.thread
         self.thread = None
 
-    def to_json(self) -> str:
-        """把请求转换成JSON字符串，可以保存起来以后再读取再请求
+    def to_dict(self) -> Dict[str, Union[int, float, str]]:
+        """把请求转换成用字符串表示的 Dict，可以保存起来以后再读取再请求
 
         Returns:
-            Json 字符串
+            Dict
         Warnings:
             这是个测试功能，不一定稳定，要配合 from_json 方法使用
             这不会保存下载状态和结果，并且 Request 的 **kwargs 参数会以简单的 json.dumps 形式保存，可能会存在问题。
@@ -152,10 +162,10 @@ class Request:
             callback 和 preparse 必须是爬虫中的方法
             downloader 和 downloader_filter 必须是某个模块中的顶级方法
         """
-        if self.thread:
+        if self.is_requesting():
             Logger.warning(
                 "The downloading state will not be retained after the request being downloaded is converted to json")
-        if self.result:
+        if self.is_finish():
             Logger.warning(
                 "The downloaded result will not be retained after the request being downloaded is converted to json")
         json_dict = {
@@ -168,46 +178,45 @@ class Request:
             'time_out': self.time_out,
             'time_out_wait': self.time_out_wait,
             'time_out_retry': self.time_out_retry,
-            'kwargs': self.kwargs,
-            'downloader':
-                (self.downloader.__module__, self.downloader.__name__),
-            'downloader_filter':
-                (self.downloader_filter.__module__, self.downloader_filter.__name__),
-            'callback': self.callback.__name__,
-            'preparse': self.preparse.__name__,
+            'save_tags': {
+                'downloader':
+                    (self.downloader.__module__, self.downloader.__name__),
+                'downloader_filter':
+                    (self.downloader_filter.__module__, self.downloader_filter.__name__),
+                'callback': self.callback.__name__,
+                'preparse': self.preparse.__name__,
+                'spider': self.spider.identity(),
+                'kwargs': self.kwargs,
+            }
         }
-        return json.dumps(json_dict)
+        return json_dict
 
     @staticmethod
-    def from_json(json_str: str, spider):
-        """ 从 json 字符串中解析出新的 Request
+    def from_dict(json_dict: Dict[str, Union[int, float, str, dict]], scheduler: 'Scheduler') -> 'Request':
+        """ 从Dict中解析出新的 Request
 
         Args:
-            json_str: json 字符串
-            spider: 负责产生这个请求的爬虫（callback 和 preparse 会从这个爬虫上获取）
+            scheduler: 需要一个调度器分配爬虫、解析函数等
+            json_dict: json 格式的 dict
 
         Returns:
             解析得到的请求
 
         Warnings:
-            这是个测试功能，不一定稳定，要配合 to_json 方法使用
-            其他警告查看 to_json 方法文档
-
+            这是个测试功能，不一定稳定，要配合 to_dict 方法使用
+            其他警告查看 to_dict 方法文档
         """
-        from requests_magic.spider import Spider
-        spider: Spider
-        result: Request
-        json_dict: dict = json.loads(json_str)
+        save_tags = json_dict['save_tags']
+        del json_dict['save_tags']
+        spider = scheduler.get_spider_by_identity(save_tags['spider'])
+        json_dict['downloader'] = getattr_in_module(*save_tags['downloader'])
+        json_dict['downloader_filter'] = getattr_in_module(*save_tags['downloader_filter'])
+        json_dict['callback'] = getattr(spider, save_tags['callback'])
+        json_dict['preparse'] = getattr(spider, save_tags['preparse'])
 
-        json_dict['downloader'] = utils.getattr_in_module(*json_dict['downloader'])
-        json_dict['downloader_filter'] = utils.getattr_in_module(*json_dict['downloader_filter'])
-        json_dict['callback'] = spider.__getattribute__(json_dict['callback'])
-        json_dict['preparse'] = spider.__getattribute__(json_dict['preparse'])
-
-        kwargs = json_dict['kwargs']
-        del json_dict['kwargs']
-
-        result = Request(**json_dict, **kwargs)
+        result: Request = Request(**json_dict)
+        result.kwargs = save_tags['kwargs']
+        result.spider = spider
         return result
 
 
@@ -253,7 +262,7 @@ class RequestThread(threading.Thread):
     def parse_downloader_exception(self, e: Exception) -> None:
         """解析调用下载器时得到的错误，用来判断是否要重试、放弃等
         """
-        retry = False
+        retry: bool = False
 
         if isinstance(e, RequestCanRetryError):
             # to retry
@@ -263,7 +272,8 @@ class RequestThread(threading.Thread):
                 message = f'{self.request} Request is time out to {self.request.url}. '
                 if self.request.time_out_retry > 0:
                     message += \
-                        f'Try again in {self.request.time_out_wait} seconds. ({self.request.time_out_retry - 1} remaining)'
+                        f'Try again in {self.request.time_out_wait} seconds.' \
+                        f' ({self.request.time_out_retry - 1} remaining)'
                 else:
                     message += 'Gave up the request'
                 Logger.warning(message)
@@ -275,7 +285,6 @@ class RequestThread(threading.Thread):
                     retry = False
             else:
                 Logger.warning(f'{self.request} Retry {self.request}')
-                retry = True
         else:
             Logger.error(f'{self.request} {e}')
             raise e