+++ requests_magic/utils.py
@@ -1,6 +1,7 @@
 """工具模块
 """
 import importlib
+from typing import Any
 
 
 def request_to_requests_kwargs(request) -> dict:
@@ -48,3 +49,16 @@ def getattr_in_module(module_name: str, func_name: str):
     """
     m = importlib.import_module(module_name)
     return getattr(m, func_name)
+
+
+class HasNameObject:
+    """ 重写了 requests magic 风格的 __str__ 方法。
+    是 '[class_name: name]' 或 '[class_name]' 这样的字符串
+    """
+
+    def __str__(self) -> str:
+        cls_name = self.__class__.__name__
+        if hasattr(self, 'name') and self.name:
+            return f'[{cls_name}: {self.name}]'
+        else:
+            return f'[{cls_name}]'